Implement SignalRingRTC.framework for iOS
This commit provides the build and implementation of the RingRTC
Framework for the iOS platform.

The RingRTC Rust code for iOS is implemented and modeled on the Android
implementation. However, the interface from Swift to and from Rust is
done through C interfaces, so none of the JNI wrapping code is needed.

On the Swift side, the C interface is Objective-C, but most of the glue
is handled by the compiler. We do use some bridging logic, especially
when interfacing with some of the WebRTC objects. Generally, the
iOS WebRTC implementation is done in Objective-C++. Since we have moved
bits of it in to Rust, we only lightly touch and adjust these RTC
interfaces.

Like the Android implementation, files in the api directory hold the
API entry points from Swift in to Rust.

Logging takes an object created in Swift (CallConnectionLogger) and
transfers ownership to Rust, which emits the logs.

There are three other types of notifications:
- Observer notifications: An object, CallConnectionObserver, is created
in Swift and ownership is transferred to Rust. The various
notifications get sent back to Swift via the object.
- Recipient notifications: An object, CallConnectionRecipient, is
created in Swift and ownership is transferred to Rust. The various
signaling notifications (i.e. to send offer, answer, etc.) are sent
back to Swift via the object.
- Direct access functions: Rust needs to invoke some construction
functions in Swift, which have dependencies on the WebRTC Objective-C++
implementation.

src/ios/SignalRingRTC is an Xcode framework project that holds the
Swift and Objective-C code for the iOS application side of RingRTC.
This is equivalent in meaning to the source at src/android. Here one
will find the application objects for CallConnectionFactory, which
is a subclass of RTCPeerConnectionFactory, CallConnection, which is
a subclass of RTCPeerConnection, and other support classes.

To make building easier and more controllable, a build script is
provided, build-ios, which is similar in responsibility to what
build-aar does for Android. We have chosen to script the Rust Cargo
builds for the rust code so that we can use 'cargo lipo' to make a
universal library for RingRTC (the Rust side), instead of using the
WebRTC gn/ninja tools.

The end result for iOS is that the following commands should be run:
  ./bin/prepare-workspace ios
  ./bin/build-ios

which will result in the output:
  ./out/SignalRingRTC.framework
  ./out/WebRTC.framework

Both of which must be included in an application for use.