Add IceReconnecting State
When receiving an IceConnectionDisconnected event, transition the FSM
to a new state, IceReconnecting, keeping track of whether the call was
connected or not at the time ICE disconnected.

Later, when receiving an IceConnected event, transition to either:

a) CallConnected, if the ICE disconnect happened after the call
connected.

OR

b) IceConnected, if the ICE disconnect happen before the call
connected.

Signed-off-by: Curt Brune <curt@signal.org>