Implement Rust FSM integration test cases
This commit adds a new cargo feature called "sim", which enables
offline testing of the call FSM.

The sim feature adds a few things:

1. A new platform backend called "sim", which sits at the same level
as the existing iOS and Android platform backends.

2. An alternative FFI backend to the to the WebRTC C++
library interface.  This allows for stubbing out the FFI calls to
WebRTC for test purposes.

The tests verify for various call event scenarios:

- the call state matches expected state
- the client event notifications are fired appropriately
- the client error notifications are fired appropriately
- signaling messages are sent appropriately

To run the integration tests, the "sim" feature must also enabled.
The details are contained within a script:

  $ cd src/rust
  $ ./scripts/run-tests

To run the tests with lots of debug logging set the |DEBUG_TESTS|
environment variable:

  $ cd src/rust
  $ DEBUG_TESTS=1 ./scripts/run-tests

For reproducing test results, the tests can be run with a specified
RANDOM_SEED:

  $ cd src/rust
  $ RANDOM_SEED=12345 ./scripts/run-tests

By default the RANDOM_SEED is the current date/time.  At the start of
the test run, the RANDOM_SEED in use is printed for re-running the
tests.

Signed-off-by: Curt Brune <curt@signal.org>