Remove non-critical error callbacks
Some messages sent through the data channel are not critical and should
not cause client errors. Such errors might terminate a call when not
really necessary.

For VideoStreamingStatus, if a send fails for some reason, it is not
critical to fail the call, so we don't report the error. The audio
part of the call at least should continue.

For Hangup, the call is already terminating. Therefore, it is not
necessary to wrap it in a failure condition.