Android: move the logging to a dedicated thread
In order to simplify the logic of connecting to (and disconnecting
from) the JVM , create a dedicated logging thread that persists for
the life of the application.  This thread connects once to the JVM and
remains connected for the life of the application.

To aid debugging, for debug builds only, also include the std::thread
ID in the log message.

Signed-off-by: Curt Brune <curt@signal.org>