Add multi-ring feature
Multi-ring allows one or more user devices to ring when called. When
one callee responds to a call (either accepting, declining, or
indicating a busy state) the caller will send an appropriate hangup to
all other devices belonging to the user.

The Calling protocol has changed, resulting in changes to the protocol
buffer for the CallMessage.

Pre-multi-ring devices will still be able to establish call sessions,
but will not be able to call linked devices.