Implement native interface
This commit adds a native interface to RingRTC. This provides native
access to more WebRTC capabilities, such as the PeerConnectionFactory,
PeerConnection, and some native media capabilities.

The native interface is used for electron-based client support, so
that WebRTC can be accessed directly with RingRTC rather than through
the Chrome backend. A script bin/build-electron is provided to build
the electron library, and a node project with a TypeScript API is
provided in src/node.

A command line interface is also available. This uses some new hooks
for simulating networks through WebRTC. This client is meant for
testing, it is not a Signal client.