Add allowMobile parameter
This change adds an allowMobile parameter to restrict calls from using mobile
data. My wife experienced a video call using 750 MB of data even though she was
on WiFI. It needs to be controllable in case you are on a metered plan since it
might otherwise make the call very expensive. My wife almost uninstalled Signal
because of this, but I got her to stay by promising I would fix it.

This change keeps the old API as wrapper methods. The wrappers can be deleted
once the corresponding changes to Signal-Android have been accepted.