Electron: Use Buffer everywhere we used to use ArrayBuffer
When converting through ArrayBuffer, keep the offset and length.

All JavaScript TypedArrays (including Buffer and Uint8ClampedArray)
wrap an underlying buffer, but they might actually be *slices* of the
underlying buffer. When we need to convert from one TypedArray to
another, we need to preserve the offset/length-in-bytes of the
original array unless we are deliberately shortening the slice.