Group ringing: don't count yourself when deciding if a call is empty
If someone rings a group call but then immediately leaves, the ring is
supposed to be cancelled (maybe it was the wrong group). However, I
had previously used the count of *users* in the call to determine this
rather than the count of *remote devices*, and the list of users
includes the local user. Switch to checking remote devices instead.

This does also mean that if you ring a group call but don't join it,
your ring will be cancelled regardless of whether there are other
people in the call. That seems acceptable, as well as not a scenario
the client apps will actually use.