Desktop: report how long ago an incoming 1:1 call was auto-ended
This is important for when a Desktop wakes up and processes a call
from five hours ago. While iOS and Android make a dedicated object to
track that call and *then* try to process the offer, Desktop does it
all in one go, and therefore they lose the timestamp of the initial
offer. Rather than trying to restructure Desktop to be more like iOS
and Android at this time, just pass through the information we have,
which is how old the offer was when we got it. That will be accurate
to within a few seconds in practice.