Electron: Use Neon's Channel to avoid polling for events/logs
Now that Neon has an API to put items on Node's task queue, we can
stop having the "observer" object poll for new events on the
JavaScript side.

This commit only avoids the polling; it still processes a queue of
recent events and a buffer of log messages. In the future, we may want
to ditch the event model entirely and enqueue direct invocations of the
various observer callbacks.