Adjust PeerConnection, VideoSinks, and RTP Data
Close PeerConnection so it's safe to drop the connection.

Avoid video sink leak by keeping VideoSinks in the PeerConnectionObserver.
This simplifies the VideoSink system by not exposing C++ VideoSinks to Rust and
instead keeps them in C++ and passing frames via the PeerConnectionObserver.
That way, the VideoSinks are destroyed when the PeerConnectionObserver is
destroyed.

Fix RTP Data to avoid deadlock.