Fire peek changed events even if the call is empty
If the client has stale data from peeking and uses a group call client
to learn new info, it won't if the call is empty unless we fire the
events that tell the client "we got new peek info, and it's empty",
which this PR does. This makes sure that that the client's call lobby
will always have the latest info (assuming the lobby uses a group call
client).