Build AARs faster
This allows the debug and release AARs to be built individually. 
Previously when you tried to build only one of them (through 
build-aar), both AARs would still be created. Now when you want to 
build only one of them, the Gradle portion of the build will take ~1/2 
the time that it previously did.

These changes also fix the bug where an AAR could be created without 
.so files in it, and the bug where the debug and release AARs would be 
exactly the same (where one overwrote the other).