Remove install-ndk-toolchains entirely in favor of `-C linker`
install-ndk-toolchains originally built toolchain directories inside
OUTPUT_DIR, which could then be used when building for Android. Last
year it got slimmed down to use a Cargo config file instead, but even
that isn't strictly necessary. We're already passing custom flags
directly to `cargo rustc`; specifying the linker is just one more.

Note that this will need to be revisited if we break apart the WebRTC
and RingRTC builds, as we've discussed doing in the past.