build-ios: Skip any dynamic linking when compiling for Catalyst
sysinfo has a "cdylib" crate-type, but Rust can't correctly compile
cdylibs on Catalyst at this time. However, we don't need the sysinfo
cdylib, because we statically link sysinfo into our own static
library libringrtc.a.

This does unfortunately mean the Catalyst build for RingRTC is never
considered clean; if that's a problem, a minimal script that touches
the output file would probably accomplish that. But it's unlikely
someone will be incrementally building for Catalyst anyway.