Add callIdFromRingId as well for iOS and Desktop
This is almost trivial on iOS, but still provided because
UInt64(ringId) does *not* do the right thing.

On Desktop, call IDs have a different representation from ring IDs in
addition to being unsigned rather than signed. We may want to change
that in the future, but in this case it will help keep the types
distinct.

Android has no unsigned integers, so call IDs and ring IDs are already
compatible there.