Node: use 'interface' instead of 'class' for PeekInfo & PeekDeviceInfo
This is TypeScript pedantry: the objects we create via Neon to pass to
'handlePeekResponse' don't come from any class, and if we added
methods to PeekInfo or PeekDeviceInfo, they wouldn't work on those
Neon-created objects. 'interface' has the semantics we want, at the
cost of not providing a constructor for testing purposes.