tests: "Synchronize" 1:1 call event queues more carefully.
A "synchronize" event normally acts as a "barrier" in a call or
connection's event queue, but during termination those event queues
are drained and shut down. If that's happened, wait for the
call/connection to finish terminating instead.

Additionally, make sure that an outgoing call's "parent" connection is
also synchronized; some tests rely on this.

This only affects tests; everything in real usage is asynchronous.