iOS: Explicitly check that the prebuild directory exists on build
This previously manifested as an opaque error for local development if
you have a file named 'prebuild-checksum' in the source directory. It
could also happen during use as a pod because the prebuilds are in one
of macOS's cache directories, which can get cleared by the system when
disk space is low.

To help keep this from happening again, also tweak
set-up-for-cocoapods to check for a build directory *first*, before
writing 'prebuild-checksum'.