android: ExceptionCheckingJNIEnv, which reports uncaught exceptions
Normally JNI handles uncaught exceptions when a native thread is
"detached" from the JVM, but RingRTC treats all its callback threads
as "daemon" threads that are only detached when the native thread
exits. ExceptionCheckingJNIEnv restores that functionality by checking
for uncaught exceptions on Drop and reporting them using the thread's
uncaught exception handler.

This is not a *recovery* mechanism; most likely the app will be
terminated in this case. But that's better than either silently
swallowing exceptions, or crashing with an opaque error the next time
the thread tries to use the JVM.