Add JoinState.PENDING, for call link calls with admin approval
During the Pending state, clients can only talk to the SFU, not to
other clients, and the set of messages the SFU will respond to is
limited (right now, only the Leave message is accepted). The local
client should still send at media keys at this stage, both to speed up
an eventual join and to provide a profile key for admins to see who
they're approving/denying.

It is possible for a client to go straight from Joining to Joined
without reporting a Pending state in between.