iOS: Collect acknowledgments for RingRTC's dependencies
This sets up a CI-enforced cargo-about generation, and then tweaks the
podspec to preserve both the RingRTC and WebRTC plists. We'll be able
to do something similar for Android and Desktop, but they may need
more post-processing, so let's start with iOS for now.

The HTML output isn't used; it's just convenient for checking what's
changed.